pyproject.toml
[tool.poetry]
name = "voice-tbot"
version = "0.1.0"
description = "A telegram bot to convert url → text → speech → audio file"
authors = ["sentenzo <kassiyskley@rambler.ru>"]
license = "MIT License"
readme = "README.md"
packages = [{include = "voice_tbot"}]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = {version = "^3.0.0b4", allow-prereleases = true}
pyttsx3 = "^2.90"
python-dotenv = "^0.21.0"
PyYAML = "^6.0"
python-box = "^6.0.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
black = "^22.8.0"
beautifulsoup4 = "^4.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
